Finished post delete

+++ controllers/postController.js
@@ -6,6 +6,8 @@ const writeImage = require(`../functions/writeImage`)
 const fs = require("fs")
 // .limit(n) returns maximum n number of items
 
+
+//! Get all posts
 const getAllPosts = asyncHandler(async (req, res) => {
   console.log(req.body)
   const posts = await Post.find().select().limit(req.query.limit).skip(req.query.skip)
@@ -13,6 +15,7 @@ const getAllPosts = asyncHandler(async (req, res) => {
   res.json({ posts, totalCount })
 })
 
+//!Create Post
 const createNewPost = asyncHandler(async (req, res) => {
   const { tags, image, title, userId, createdBy, text } = req.body
   if (tags.length === 0 || !title || !text || !userId || !createdBy) {
@@ -54,46 +57,17 @@ const createNewPost = asyncHandler(async (req, res) => {
   const postObject = { tags, image: formatedImage, title, userId, createdBy, text }
   const postResponse = await Post.create(postObject)
   if (postResponse) {
-    res.status(201).json(postResponse)
+    res.status(201).json({postResponse, message:{text:`Post created`, type:`confirm`}})
   } else {
-    res.status(400).json({ message: `Invalid data` })
+    res.status(400).json({ message:{text:`Invalid data`, type:`error`} })
   }
 })
 
+
+
+//! Update post
+//After profile
 const updatePost = asyncHandler(async (req, res) => {
-  // console.log(req.body)
-
-  // Sending a comment
-  if (req.body?.updateType === `comment`) {
-    const { comment, commenterUsername, commenterId, postId, commentId } = req.body
-    if (!comment || !commenterUsername || !commenterId || !postId || !commentId) {
-      return res.status(400).json({ message: `All fields are required` })
-    }
-
-    const commentedPost = await Post.findById(postId).exec()
-    if (!commentedPost) {
-      return res.status(400).json({ message: `Post not found` })
-    }
-    const date = new Date()
-    let day = date.getDate() < 10 ? `0${date.getDate()}` : date.getDate()
-    let month = date.getMonth() + 1 < 10 ? `0${date.getMonth() + 1}` : date.getMonth() + 1
-    let year = date.getFullYear()
-    let currentDate = `${day}-${month}-${year}`
-    const commentObejct = {
-      commentText: comment,
-      commenterUsername,
-      commenterId:Number(commenterId),
-      commentPosted: currentDate,
-      postId,
-      commentId,
-    }
-    commentedPost.comments.unshift(commentObejct)
-    console.log(commentObejct)
-    const commentSent = await commentedPost.save()
-    res.json({message:{text:`Comment posted`, type:`confirm`}, commentObejct})
-    return
-  }
-
   const { editedPost, postId, userId } = req.body
   if (!editedPost || !userId) {
     res.status(400).json({ message: `Error, no data recieved` })
@@ -113,25 +87,25 @@ const updatePost = asyncHandler(async (req, res) => {
   res.json(updatePost)
 })
 
+
+//!Delete post
+//!Finished
 const deletePost = asyncHandler(async (req, res) => {
-  const { postId, userId, deleteType } = req.body
-  // console.log(req.body)
-  if (!postId || !userId || !deleteType) {
-    return res.status(400).json({ message: `No ID recieved` })
+  const { postId, userId } = req.body
+  if (!postId || !userId) {
+    return res.status(400).json({ message: {text: `No ID recieved`, type:`error`} })
   }
     const post = await Post.findById(postId).exec()
     if (!post) {
-      return res.status(400).json({ message: `Post not found` })
+      return res.status(400).json({ message: {text:`Post not found`, type:`error`} })
     } else if (post && post.userId !== userId) {
-      return res.status(400).json({ message: `Acces denied` })
+      return res.status(400).json({ message: {text: `Access denied`, type:`error`} })
     }
     const result = await post.deleteOne()
     const comments = await Comment.deleteMany({postId:postId})
 
-    fs.unlinkSync("public/images/" + post.image)
-    console.log(result)
-    const reply = `Post ${result.title} has been deleted`
-    res.json(reply)
+    post.image && fs.unlinkSync("public/images/" + post.image)
+    res.json({message:{text:`Post deleted`, type:`confirm`}})
